+++ lib/inertia/plug.ex
@@ -54,6 +54,8 @@ defmodule Inertia.Plug do
         |> put_private(:inertia_request, true)
         |> detect_partial_reload()
         |> detect_reset()
+        |> detect_except_once_props()
+        |> detect_scroll_merge_intent()
         |> convert_redirects()
         |> check_version()
 
@@ -86,6 +88,29 @@ defmodule Inertia.Plug do
       end
 
     put_private(conn, :inertia_reset, resets)
+  end
+
+  defp detect_except_once_props(conn) do
+    except_once =
+      case get_req_header(conn, "x-inertia-except-once-props") do
+        [stringified_list] when is_binary(stringified_list) ->
+          String.split(stringified_list, ",")
+
+        _ ->
+          []
+      end
+
+    put_private(conn, :inertia_except_once_props, except_once)
+  end
+
+  defp detect_scroll_merge_intent(conn) do
+    intent =
+      case get_req_header(conn, "x-inertia-infinite-scroll-merge-intent") do
+        [intent] when intent in ["append", "prepend"] -> intent
+        _ -> "append"
+      end
+
+    put_private(conn, :inertia_scroll_merge_intent, intent)
   end
 
   defp get_partial_only(conn) do