+++ test/support/my_app/lib/my_app_web/router.ex
@@ -38,6 +38,18 @@ defmodule MyAppWeb.Router do
     get "/preserved_case_props", PageController, :preserved_case_props
     get "/local_ssr", PageController, :local_ssr
     get "/force_redirect", PageController, :force_redirect
+    get "/once_props", PageController, :once_props
+    get "/once_props_fresh", PageController, :once_props_fresh
+    get "/once_props_with_expiration", PageController, :once_props_with_expiration
+    get "/once_props_with_custom_key", PageController, :once_props_with_custom_key
+    get "/once_props_camelized", PageController, :once_props_camelized
+    get "/once_props_with_deferred", PageController, :once_props_with_deferred
+    get "/scroll_props", PageController, :scroll_props
+    get "/scroll_props_with_custom_wrapper", PageController, :scroll_props_with_custom_wrapper
+    get "/scroll_props_with_custom_page_name", PageController, :scroll_props_with_custom_page_name
+    get "/scroll_props_lazy", PageController, :scroll_props_lazy
+    get "/scroll_props_camelized", PageController, :scroll_props_camelized
+    get "/scroll_props_with_custom_metadata", PageController, :scroll_props_with_custom_metadata
     put "/", PageController, :update
     patch "/", PageController, :patch
     delete "/", PageController, :delete